fix crashes, including 1. a multithreading crash due to wrong mutex use 2. a crash due to multiple calls to FT_FreeType_Done

+++ GCanvas/core/src/gcanvas/GCanvas2dContext.h
@@ -129,7 +129,7 @@ public:
     GColorRGBA mStrokeColor;
 
     GColorRGBA mShadowColor;
-    int mShadowBlur;
+    float mShadowBlur;
     float mShadowOffsetX;
     float mShadowOffsetY;
